add rule and submit page
